Add BLE reconnect detection, code review cleanup
IOHIDManager watcher detects Logitech HID device arrival/removal
via CFRunLoop, triggering reconnect to re-divert buttons after BLE
drops. Keepalive checks diversion state every 5 minutes as a safety
net for silent BLE reconnects where the HID handle stays valid but
firmware state resets.

Code review fixes:
- Extract connect_device(), execute_and_notify(), format_hex(),
  classify_error(), action_description() to eliminate duplication
- ActionOutcome enum replaces ambiguous bool return from execute()
- Move SAMPLE_CONFIG to config.rs, add ACCESSIBILITY_ERROR const
- Fix run_listen_only with proper backoff and error deduplication
- Deduplicate FFI declarations in platform.rs, fix c-string literals
- Collapse permission error handling in main.rs
- Fix service.rs plist label to match bundle ID (com.jlevere.hidpp)
- Fix RESPONSE_TIMEOUT_MS type from i32 to u64

+++ crates/hidpp-transport/src/native.rs
@@ -17,7 +17,8 @@ const HIDPP_USAGE_PAGE_FF43: u16 = 0xFF43;
 const HIDPP_USAGE_PAGE_FF00: u16 = 0xFF00;
 
 /// Default timeout for waiting on a response (ms).
-const RESPONSE_TIMEOUT_MS: i32 = 2000;
+/// Used both as hidapi read timeout (i32) and tokio timeout (u64).
+const RESPONSE_TIMEOUT_MS: u64 = 2000;
 
 /// Enumerate connected Logitech HID++ devices.
 pub struct HidapiEnumerator {
@@ -249,7 +250,7 @@ impl HidapiTransport {
 
         // Wait for response with timeout.
         match tokio::time::timeout(
-            Duration::from_millis(RESPONSE_TIMEOUT_MS as u64),
+            Duration::from_millis(RESPONSE_TIMEOUT_MS),
             response_rx,
         )
         .await
